refactor: useReplyComplete 훅 분리
