+++ src/pages/ReviewReplyPage/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useHandleReceiver } from './useHandleReceiver'
 export { default as useHandleQuestion } from './useHandleQuestion'
 export { default as useClickNextButton } from './useClickNextButton'
+export { default as useReplyComplete } from './useReplyComplete'
